Create ch3_prog_5_2d_matrix_set_operations_1.py
Objective: Construct matrix converters and transforms for scaling of 2D objects. 
The operators should be expressed as functions and sets of objects processed as a single entity. 

Comments: The original data must be expressed as vertices which can then be  converted into matrices (by the Numpy module). 
A scaling matrix is created. Simple matrix multiplication will then achieve the amplification but the result is another matrix which must finally be flattened into a simple x-y list of coordinates that tkinter "create_line()" function will understand.