Create ch4_3d_shape_convert_2Dto4D_1.py
Objective: Convert a 2D planar shape into a list of 4D vertex vectors. 

Comments: 2d shape vertices are converted to lists of 4d (homogeneous) matrix/vectors. 
Each vector is the location of a vertex in 3d (homogeneous) space. 

A 'shape' is a 2d planar line of arbtrary number of segments. 
The final list corresponds to a shape object in 3 dimensions. 
The shape object is a Python list of 4D vertices. 
"canonical" (definition): A mathematical expression in the simplest or standard form.