Create ch4_prog_5_3d_shape_set_2Dto4D_rotation_1.py
Objective: Convert a 2D planar shape set into a list of 4D vertex vectors, 
and then perform transform geometry (scaling, translation and rotation). 

Comments: 2d shape vertices are converted to lists of 4d (homogeneous) matrix/vectors. 
Each vector is the location of a vertex in 3d (homogeneous) space. 

A 'shape' is a 2d planar line of arbtrary number of segments. 
A 'set' is a collection of separate shapes. 

The final list "fish_set" corresponds to a set of shape objects in 3 dimensions. 
The shape object is a list of 4D vertices.