Create ch9_prog_5_opengl_boxy_triangles_1.py
Objective: Use variable parameters to define multiple objects. 

Explanation: The idea is to be able pre-construct a shape like a box and then be able to draw multiple instances each with their own 
shape, position and rotation dynamics. 
Variable parameters rather than constants are used to populate vertex arrays. 
The names of the variables are used as OpenGL vertex arrays. 
That is, that are referenced by their Python list names. 
 This avoids the awkwardness and some of the speed performance penalty associated with the "glBegin ... glEnd" style of 3D object definition. 

Three different animation functions are available: 
1. glutIdleFunc(all_displays)     
2. glutDisplayFunc(all_displays)   
3. glutTimerFunc(1,update,0) calling "update" which calls "all_displays()".