Create ch10_prog_5_intersect.py
Objective: Locate the point of intersection of two lines in 3D space.

Comments: The lines have to be coplanar. No test is done to verify that the lines are coplanar. Since this is designed only to be used to work with lines in the plane of the same triangle, the test is not necessary.